Update Neovim configuration
This commit introduces several changes to the Neovim configuration.
Firstly, it adds new keybindings for chat and tab operations.
Secondly, it includes the setup for the CopilotChat plugin with
custom prompts for various actions like Explain, Review, Tests, Fix,
Optimize, Docs, FixDiagnostic, Commit, and CommitStaged.
It also modifies the clipboard setting to include 'unnamedplus'.
Finally, it introduces new user commands for interacting with
the CopilotChat plugin and for converting text to PascalCase.